pyproject.toml
[project]
name = "sipit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "fire>=0.7.1",
    "hf-transfer>=0.1.9",
    "jupyter>=1.1.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.3",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]
